requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Mail>=0.9.1
qrcode>=7.3
Pillow>=9.0.0
python-dotenv>=0.19.0
validate-docbr>=1.10.0
gunicorn>=20.1.0
